+++ src/builtins/core/iter.h
@@ -1,0 +1,10 @@
+#ifndef ITER_H
+#define ITER_H
+
+#include "jstar.h"
+
+// class Iterable
+JSR_NATIVE(jsr_core_iter_Iterable_join);
+// end
+
+#endif
